+++ .gitignore
@@ -1,5 +1,6 @@
 train.csv
 test.csv
+submission.csv
 sample_submit.csv
 
 __pycache__/
